+++ src/EllaX.Clients/IBlockchainClient.cs
@@ -8,7 +8,7 @@ namespace EllaX.Clients
     public interface IBlockchainClient
     {
         Web3 Web3 { get; }
-        Task<Block> GetBlockWithTransactionsAsync(uint blockNumber, CancellationToken cancellationToken = default);
-        Task<uint> GetLatestBlockNumberAsync(CancellationToken cancellationToken = default);
+        Task<Block> GetBlockWithTransactionsAsync(ulong blockNumber, CancellationToken cancellationToken = default);
+        Task<ulong> GetLatestBlockNumberAsync(CancellationToken cancellationToken = default);
     }
 }
